# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: API de Pagamentos Sindireceita
  version: "1.0.0"
  summary: Cobrança e conciliação (PIX, boleto, checkout, DDA) sobre bancos do cliente.
  description: |
    API REST `/v1` que empresas usam como meio de cobrança sobre a **sua própria**
    credencial bancária (modelo PSP-Indireto — o dinheiro liquida direto na conta
    da empresa). Primeiro banco em produção: **C6**.

    Guia de integração/onboarding: `integration-guide.md` (mesma pasta).

    Convenções globais:
    - **Autenticação:** `Authorization: Bearer <token>` em toda rota `/v1`. A
      empresa-cliente é derivada do token, nunca de um parâmetro. Token
      ausente/inválido → `401`.
    - **Valores monetários:** sempre em **centavos** (inteiro). R$ 10,00 → `1000`.
    - **Datas/hora:** RFC 3339 (ex. `2026-08-20T00:00:00-03:00`), exceto a janela
      do extrato que usa `YYYY-MM-DD`.
    - **Idempotência:** rotas de escrita exigem o header `Idempotency-Key`.
    - **Erros:** envelope uniforme `{ "error": "..." }` (não é RFC 7807; ver guia).
    - **Limites de entrada:** corpo até 1 MiB (`413` acima); campos JSON
      desconhecidos são rejeitados (`400`).
    - **Rate limit:** token-bucket por empresa (burst 20, 10 req/s) → `429`.
  contact:
    name: Suporte técnico Sindireceita
  license:
    name: Proprietária
servers:
  - url: https://api.pagamentos.sindireceita.example
    description: Produção (URL real definida por ambiente)
...
    description: Homologação
security:
  - bearerAuth: []
...
paths:
...
  /v1/charges/{id}:
    get:
      operationId: getCharge
      tags: [Cobranças]
      summary: Consultar cobrança
      parameters:
        - $ref: "#/components/parameters/PathId"
      responses:
        "200":
          description: Cobrança.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Charge" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
        "500": { $ref: "#/components/responses/InternalError" }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: Token de API escopado por empresa-cliente.
  parameters:
...
    PathId:
      name: id
      in: path
      required: true
      schema: { type: string }
...
  responses:
...
    Unauthorized:
      description: Token ausente ou inválido.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
          example: { error: "unauthorized" }
    NotFound:
      description: Recurso inexistente ou pertencente a outra empresa (mesma resposta, para não vazar existência).
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
          example: { error: "not found" }
...
    InternalError:
      description: Erro interno (transitório).
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
          example: { error: "internal error" }
  schemas:
    Error:
      type: object
      description: Envelope de erro uniforme (não é RFC 7807).
      required: [error]
      properties:
        error: { type: string, description: Mensagem curta e segura (sem detalhe interno). }
...
    Charge:
      type: object
      properties:
        id: { type: string }
        tenant_id: { type: string }
        endpoint: { type: string }
        amount_cents: { type: integer, format: int64 }
        currency: { type: string }
        status: { type: string }
        tx_id: { type: string }
...